# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: API Integration
# during dev, should point to your local machine
host: localhost:5072
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /api/jobInvites:
    # binds a127 app logic to a route
    x-swagger-router-controller: invites_ctrl
    post:
      tags: ['Invites Module']
      description: Add New Job Invites
      summary:  Add New Job Invites
      # used as the method name of the controller
      operationId: jobInvites
      parameters:
        - name: body
          in: body
          description:  Add New Job Invites
          schema:
           $ref: '#/definitions/jobInvites'
          required: true       
        - name: Authorization
          in: header
          description: Authorization
          required: true
          type: string               
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/acceptJobInvite:
    # binds a127 app logic to a route
    x-swagger-router-controller: invites_ctrl
    post:
      tags: ['Invites Module']
      description: Accept  Job Invites
      summary:  Accept Job Invites
      # used as the method name of the controller
      operationId: acceptJobInvite
      parameters:
        - name: body
          in: body
          description: Accept Job Invites
          schema:
           $ref: '#/definitions/acceptInvite'
          required: true       
        - name: Authorization
          in: header
          description: Authorization
          required: true
          type: string               
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/shareJobAcceptence:
    # binds a127 app logic to a route
    x-swagger-router-controller: invites_ctrl
    post:
      tags: ['Invites Module']
      description: Assign Share job to supervisor
      summary:  Assign Share job to supervisor
      # used as the method name of the controller
      operationId: shareJobAcceptence
      parameters:
        - name: body
          in: body
          description: Assign Share job to supervisor
          schema:
           $ref: '#/definitions/shareJobAcceptences'
          required: true       
        - name: Authorization
          in: header
          description: Authorization
          required: true
          type: string               
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/rejectJobInvite:
    # binds a127 app logic to a route
    x-swagger-router-controller: invites_ctrl
    post:
      tags: ['Invites Module']
      description: Reject  Job Invites
      summary:  Reject Job Invites
      # used as the method name of the controller
      operationId: rejectJobInvite
      parameters:
        - name: body
          in: body
          description: Reject Job Invites
          schema:
           $ref: '#/definitions/rejectInvite'
          required: true       
        - name: Authorization
          in: header
          description: Authorization
          required: true
          type: string               
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /api/getInvitiesList:
    # binds a127 app logic to a route
    x-swagger-router-controller: invites_ctrl
    post:
      tags: ['Invites Module']
      description: Get Invities List
      summary: Get Invities Job List
      # used as the method name of the controller
      operationId: getInvitiesList
      parameters:
        - name: body
          in: body
          description: get Invities By Contractor Id
          schema:
           $ref: '#/definitions/getInvitesList'
          required: true           
        - name: Authorization
          in: header
          description: Authorization
          required: true
          type: string                          
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/getjobInvites:
    # binds a127 app logic to a route
    x-swagger-router-controller: invites_ctrl
    post:
      tags: ['Invites Module']
      description: Get Invities List
      summary: Get Invities Job List
      # used as the method name of the controller
      operationId: getjobInvites
      parameters:
        - name: body
          in: body
          description: get Invities By Contractor Id
          schema:
           $ref: '#/definitions/getInvitesList'
          required: true           
        - name: Authorization
          in: header
          description: Authorization
          required: true
          type: string                          
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"         
...
  /api/getInvite_Supervisor_Id:
    # binds a127 app logic to a route
    x-swagger-router-controller: invites_ctrl
    post:
      tags: ['Invites Module']
      description: Get supervisor assign job
      summary:  Get supervisor assign job
      # used as the method name of the controller
      operationId: getInvite_Supervisor_Id
      parameters:
        - name: body
          in: body
          description: Get supervisor assign job
          schema:
           $ref: '#/definitions/supervisorjob'
          required: true           
        - name: Authorization
          in: header
          description: Authorization
          required: true
          type: string                          
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  supervisorjob:
    type: object
    properties:
      job_id: 
        type: string
      approval_status:
        type: boolean
      invities_status:
        type: boolean
...
  getInvitesList:
    type: object
    properties:
      page:
        type: string         
      count:
        type: string
      users_id:
        type: string
...
  jobInvites:
    type: object
    properties:
      invities_assign_by: 
         type: string
      invities_assign_to:
         type: string
      job_id: 
         type: string 
  acceptInvite:
    type: object
    properties:
      job_id:
         type: string
      status:
         type: string
      merge_status:
         type: string
      approval_status:
         type: string
      job_assign_to:
         type: string
...
  shareJobAcceptences:
    type: object
    properties:
      job_id:
         type: string
      approval_status:
         type: string
      job_assign_to:
         type: string
  rejectInvite:
    type: object
    properties:
      status: 
         type: string
...
  appResponse:
    required:
      - code
    properties:
      code:
        type: number
      status:
        type: string        
      message:
        type: string        
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string